Change also second and third argument of extirpolate! in-place
Contrary to initial expectations, `X` and `Y` arguments of
`extirpolate!` are not `Range`s, instead they are vectors built inside
`trig_sum!` and not needed after they are passed to `extirpolate!`.
Thus, they can be safely changed in-place in order to save some bits of
memory.

Update tests taking care of the fact that the second and third arguments
are changed in-place, so they cannot be `Range`s.